+++ lib/ask/rails/railtie.rb
@@ -3,9 +3,6 @@
 module Ask
   module Rails
     class Railtie < ::Rails::Railtie
-      generators do
-        require_relative "generators/install/install_generator"
-      end
     end
   end
 end